Rename publish/ to its package name
