+++ scripts/configs/clean.js
@@ -3,6 +3,6 @@ module.exports = {
     amd: ['temp'],
     dist: ['dist/js/*', 'dist/svg/*', 'dist/css/*'],
     app: ['app/dist/css/*'],
-    publish: ['publish/dist', 'publish/sass']
+    publish: ['ids-enterprise/dist', 'ids-enterprise/sass']
   }
 };
